it is hard to use a unified code to adapt the SpinalConfig, so clone them into each package.
