+++ tester/src/test/scala/spinal/tester/scalatest/package.scala
@@ -3,8 +3,25 @@ package spinal.tester
 import spinal.core._
 
 package object scalatest {
-  import spinal.tester.SpinalConfig
   val simWorkspacePath = spinal.tester.simWorkspacePath
+  
+  def SpinalConfig(
+      mode: SpinalMode = null,
+      dumpWave: DumpWaveConfig = null,
+      defaultConfigForClockDomains: ClockDomainConfig = ClockDomainConfig(),
+      defaultClockDomainFrequency: IClockDomainFrequency = UnknownFrequency(),
+      allowOutOfRangeLiterals: Boolean = false,
+      verbose: Boolean = false,
+      targetDirectory: String = simWorkspacePath
+  ) = spinal.core.SpinalConfig(
+    mode = mode,
+    dumpWave = dumpWave,
+    targetDirectory = targetDirectory,
+    allowOutOfRangeLiterals = allowOutOfRangeLiterals,
+    verbose = verbose,
+    defaultConfigForClockDomains = defaultConfigForClockDomains,
+    defaultClockDomainFrequency = defaultClockDomainFrequency
+  )
 
   def SpinalVerilog[T <: Component](gen: => T): SpinalReport[T] = SpinalConfig().generateVerilog(gen)
   def SpinalVhdl[T <: Component](gen: => T): SpinalReport[T] = SpinalConfig().generateVerilog(gen)
